Added payment with paypal

+++ RosenW-OnlineShop/onlineshop/routes/index.js
@@ -7,6 +7,68 @@ let router = express.Router();
 let Recaptcha = require('express-recaptcha').Recaptcha;
 let $ = require('jquery');
 let request = require('request');
+let braintree = require('braintree');
+let Printer = require('node-printer');
+let bixolon = new Printer('BIXOLON-SRP-350II');
+
+//credit card payment keys
+let merchId = '9mjmz4gm33rrmbd2';
+let publicKey = 'yy9fyqg8m8yqdrhs';
+let privateKey = '955e3451756ce5f6ab95eb47ce159245';
+
+let gateway = braintree.connect({
+  environment: braintree.Environment.Sandbox,
+  merchantId: merchId,
+  publicKey: publicKey,
+  privateKey: privateKey
+});
+
+gateway.config.timeout = 10000;
+
+//set up merchant account
+
+merchantAccountParams = {
+  individual: {
+    firstName: "Jane",
+    lastName: "Doe",
+    email: "jane@14ladders.com",
+    phone: "5553334444",
+    dateOfBirth: "1981-11-19",
+    ssn: "456-45-4567",
+    address: {
+      streetAddress: "111 Main St",
+      locality: "Chicago",
+      region: "IL",
+      postalCode: "60622"
+    }
+  },
+  business: {
+    legalName: "Jane's Ladders",
+    dbaName: "Jane's Ladders",
+    taxId: "98-7654321",
+    address: {
+      streetAddress: "111 Main St",
+      locality: "Chicago",
+      region: "IL",
+      postalCode: "60622"
+    }
+  },
+  funding: {
+    descriptor: "Blue Ladders",
+    destination: braintree.MerchantAccount.FundingDestination.Bank,
+    email: "mailsender6000@gmail.com",
+    mobilePhone: "5555555555",
+    accountNumber: "1123581321",
+    routingNumber: "071101307"
+  },
+  tosAccepted: true,
+  masterMerchantAccountId: "14ladders_marketplace",
+  id: "blue_ladders_store"
+};
+
+gateway.merchantAccount.create(merchantAccountParams, function (err, result) {
+});
+
 // set up email
 let transporter = nodemailer.createTransport({
   service: 'gmail',
@@ -248,7 +310,7 @@ router.get('/check', function(req, res, next) {
   }
   let purchases = [];
   let count = 1;
-  client.query(sqlFormatter("select p.date, p.id, a.first_name, st.name as state from accounts as a join purchases as p on a.id = p.userid join states as st on st.id = p.state order by state asc, p.date asc"))
+  client.query(sqlFormatter("select p.date, p.id, a.first_name, a.address, st.name as state from accounts as a join purchases as p on a.id = p.userid join states as st on st.id = p.state order by state asc, p.date asc"))
   .then((data)=>{
     data.rows.forEach((row)=>{
       row.number = count++;
@@ -483,12 +545,18 @@ router.get('/buy', function(req, res){
   if(isAdmin || !isLoggedIn){
     res.redirect('/');
   }
-  res.render('buy', {data:{'isLoggedIn': isLoggedIn,'user': username, 'isAdmin': isAdmin}});
+  let wd = req.query.wd;
+  gateway.clientToken.generate({}, function (err, response) {
+    let clientToken = response.clientToken
+    res.render('buy', {data:{'isLoggedIn': isLoggedIn,'user': username, 'isAdmin': isAdmin, 'wd': wd, 'ct': clientToken}});
+  });
 });
 
 /* POST buy page */
 router.post('/buy', function(req, res){
   let pass = true;
+  //get nonce
+  let nonce = req.body.nonce;
   //get user prods
   client
   .query(sqlFormatter("select pr.id, ci.quantity, pr.quantity as max from cart_items as ci join products as pr on ci.prodid = pr.id join shopping_carts as sc on ci.cartid = sc.id where sc.userid = %L;", loggedInUserId))
@@ -504,28 +572,39 @@ router.post('/buy', function(req, res){
     });
 
     if(pass){
-      data.rows.forEach((row)=>{
-        //row.max = row.max - row.quant
-        let newQuantity = row.max - row.quantity;
-        client.query(sqlFormatter("update products set quantity = %L where id = %L", newQuantity, row.id));
-      });
-      //make a purchase
-      let currentPurchaseId = generateId();
-      client.query(sqlFormatter("insert into purchases values(%L, %L, 0, %L)", currentPurchaseId, loggedInUserId, getDate()));
+      //make test credit card transaction
+      gateway.transaction.sale({
+        amount: "10.00",
+        paymentMethodNonce: nonce,
+        options: {
+          submitForSettlement: true
+        }
+      }, function (err, result) {
+        pass = result.success;
+        console.log(pass);
+        if(result.success){
+          data.rows.forEach((row)=>{
+            //row.max = row.max - row.quant
+            let newQuantity = row.max - row.quantity;
+            client.query(sqlFormatter("update products set quantity = %L where id = %L", newQuantity, row.id));
+          });
 
-      data.rows.forEach((row)=>{
-        //add products to purchase
-        client.query(sqlFormatter("insert into purchase_items values(%L, %L, %L, %L)", generateId(), row.id, currentPurchaseId, row.quantity));
+          //make a purchase
+          let currentPurchaseId = generateId();
+          client.query(sqlFormatter("insert into purchases values(%L, %L, 0, %L)", currentPurchaseId, loggedInUserId, getDate()));
+    
+          data.rows.forEach((row)=>{
+            //add products to purchase
+            client.query(sqlFormatter("insert into purchase_items values(%L, %L, %L, %L)", generateId(), row.id, currentPurchaseId, row.quantity));
+          });
+
+          //cart clean up
+          client.query(sqlFormatter("delete from cart_items as ci using shopping_carts as sc where ci.cartid = sc.id and sc.userid = %L", loggedInUserId))
+          .then(res.redirect(303, '/orders'));  
+        }else{
+          res.redirect(303, '/buy?wd=1');
+        }
       });
-    }
-  })
-  .then(()=>{
-    if(pass){
-      //cart clean up
-      client.query(sqlFormatter("delete from cart_items as ci using shopping_carts as sc where ci.cartid = sc.id and sc.userid = %L", loggedInUserId))
-      .then(res.redirect(303, '/'));  
-    }else{
-      res.redirect(303, '/orders');
     }
   });
 });
